Finish the first version of prepare.py for howto100M youtube subs
